As demoed.  tap_schema.db contains only gaia_dr3.gaia
gaia_dr3.gaia is cloned from gaia_dr3_source at Caltech.

The note about filters: my understanding is that `filters=`
should take a list of lists, to handle multiple conditions, but
passing them that way produces an error message.  It might be
necessary to do the `[0]` if there is only one condition.